Adds mapper to dto to all entities. Creates separate mappers for tasks, reminders, and events
